A second round of Ruby linting
So I gave in and made the changes…

I also replaced the backticks in line 32 with a system call as I am
getting some weirdness with them as of a Ruby update this morning.
Might be unrelated though.

Please also test the Fusion changes from this and my previous pull
request as I don’t have it on hand to test with.