Improve purge playbook
We run a zap twice here since we can't combine with_items and
with_sequence yet. This will land in Ansible 2.0. We destroy all the
partitions and re-create new labels as well. We do this twice just
because sometimes one is not sufficient. We also call partprobe to make
sure that the kernel sees the changes.

Signed-off-by: leseb <seb@redhat.com>