ceph-osd: Change ownership of osd directories.
When using the 4th scenario in ceph-osd role (directory instead of a disk for osd), the directories given in the `osd_directories` variable in `group_vars/osds` are created with `root:root` ownership.
This cause the command `ceph-disk activate` to fail in the task 'activate OSD(s)' in `roles/ceph-osd/tasks/scenarios/osd_directory.yml` because of
the arguments `--setuser ceph --setgroup ceph` that are passed to the command `/usr/bin/ceph-osd` launched by `ceph-disk activate`.