Ensure purge-cluster doesn't call install tasks
Currently calling purge-cluster calls ceph-common role
which ends up installing common ceph deps. This isn't
correct as we are in the process of purging the cluster.

Add a new purge_cluster var and ensure all the tasks tagged
with package-install are skipped if purge_cluster is true.
Default value of purge_cluster is false.

A better way to fix this would be to use --skip-tags, but
unfortunately there is no way to embed skip-tags inside a
playbook, thus falling back to this method.

Signed-off-by: Deepak C Shetty <deepakcs@redhat.com>