choose_disk: Converting storage units during size comparison
When comparing disk's size with lt() or gt() we face a situation where sizes are
expressed in various units and cases.

We'd like being able to ask for a disk greater than 500 MB even if its own size is
reported in TB.

A typical example looks like :
    ansible_devices = {'sr0': {'size': '1.00 TB'}}
    disk_0 = {'storage_disks': {'size': 'gt(500.00 MB)'}}

This commit does convert all the known storage units in bytes, therefore the
gt() and lt() comparaison are easier to perform.

Signed-off-by: Erwan Velu <erwan@redhat.com>