choose_disk: Adding support for legacy disks naming
The new way to describe disks we are looking for, could be confusing for some.
To make a smoother transition and to get workarounds for the new detection code,
this patch introduce the support of the classical/legacy/suxy naming.

It is now possible to use the "legacy_disks" variable under the traditional string,
space separated list of block devices to be used.

This code path disable the consistency disk name renaming but keep the matching
code to detect a missing disk prior calling ceph-disk.

Disks listed in the legacy_disks, are transformed into a dict like :

    legacy_lookup_disks: "/dev/sda /dev/sdb /dev/sdc"
->
    {'legacy_0': {'bdev': '/dev/sda'}, 'legacy_1': {'bdev': '/dev/sdb'}, 'legacy_2': {'bdev': '/dev/sdc'}}

The "bdev" aka "block dev" entry represent the path of a given disk.
This information is populated when processing the 'legacy_disks' entry.

Bdev entries are also populated to devices reported in ansible_devices when detecting the persistent name.
In that case, we only keep the persistent bdev name and not the original one.