choose_disk: Returning "devices" structure
Once the processing is over, the found devices should be returned as facts.
The disks structure is returned over the "devices" facts.

Note that if all detected devices are enabled for ceph, the "changed" parameter
will be returned as false to indicate that no changed occured and no need to do
any further configuration.

Signed-off-by: Erwan Velu <erwan@redhat.com>