choose_disk: pre-sorting ceph-disks when looking for physical disks
When searching for a match between a lookup disk and the physical disks, we
shall first check against the existing disks.

The broken scenario was :
- system have 2 disks already configured
- system have been added disks
- system have now 20 disks
- lookup expect 2 disks

In that configuration, the module was reporting the two first devices of the
sorted system list while 2 were already configured. As a result, two new disks
we provisionned.

This patch avoid this situation by creating a list of sorted devices where the
first disks of the list are the ceph one.

In the same scenario, looking for 2 disks over 20, puts the 2 already configured
disks first making the match being performed against them.

As a result, the module reports no changes with changed=false.

Signed-off-by: Erwan Velu <erwan@redhat.com>