Remove reference to old OSD docker extra env vars
This variable is already defined as a global default in the OSD role and
was not being kept in sync as we now require the '-e' parameter prefixed
to each variable. This was also missing the CLUSTER environment variable
that is defined in the global default version of
ceph_osd_docker_extra_env.