mon: fix with_items being evaluated before when
Ansible evaluates the 'with_items' before the 'when' so if the inventory
does not have the group declared it'll fail. To fix this, we set an
empty array to make the with_items happy and then evaluate with the
'when'.

Signed-off-by: Sébastien Han <seb@redhat.com>
(cherry picked from commit 05331a26348e162e4083131d2a259f6c983f3ec9)
Signed-off-by: Sébastien Han <seb@redhat.com>