Drop ceph-iscsi-gw role and all references
This role is an empty placeholder and it creates a conflict with the
ceph-iscsi-ansible role. The latter was going to merged into this repo
months ago, but it hasn't, thus, remove the empty shell here and thus
resolve the conflict.

Consequently, also remove references to the iscsi_gw_vms group, and
all iSCSI-related cruft in Vagrant files.

Fixes #1031.

Signed-off-by: Florian Haas <florian@hastexo.com>