Tests: get directories names shorter with symlinks
Add a script to generate links in `./tests/` with a the first 7 chars of the
directories' names sha1 as file name. This is to avoid a bug in libvirt
with the path length.

https://bugzilla.redhat.com/show_bug.cgi?id=1324587
eg:

`failed: internal error: Monitor path /var/lib/libvirt/qemu/domain-bs-docker-cl
uster-dmcrypt-journal-collocation_mon0_1499294943_ba9faf7bf296533177f6/monitor.
sock too big for destination`

As we can't upgrade libvirt in our CI for some reason, we need to get the
directories name shorter in order to workaround this issue.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>