rpm: better comments for file removals
As of 54d7a81241eac26d87e2bee513df5efd8866a586, we're doing more than
stripping CoreOS files in the RPM now. Move the comments around to
better match the code that does what the comments describe.

The purpose of this change is to make it easier to read this part of the
RPM spec file.