Restart service based on host specific vars
The current approach to restarting rgw/mds servers can result in
inconsistent behaviour. Sometimes this will attempt to restart the
initial hosts "ansible_hostname" service on the delegated_to host, which
will fail.

This PR changes to specifically use the delegate_to: host's hostvars.
This will prevent a situation where the incorrect hostvars are used.

This doesn't apply to master, since the approach to restart handlers has
moved to in role and no longer runs as a "run_once" with delegation.

Fixes: #1921