Auto discovery of devices for collocated and non-collocated scenarios
The changes in this pull request were tested and deployed into production
in our environment.  The collocated and non-collocated scenarios are the ones
we focused on.

Added the ceph-osd/tasks/discover_devices.yml to inventory the ansible_devices
dictionary for the list of usable devices.  The discovered devices in the
devices list are applicable to all the supported scenarios.

In the ceph-osd/tasks/main.yml, include the discover_devices.yml before the
check_devices.yml to populate the devices and dedicated_devices (when
non-collocated scenario is chosen) lists.  Use the osd_scenario to include
the correct scenario in tasks/scenarios/*.yml.

Added additional checks in the check_mandatory_vars.yml file.

The directive always_run is obsolete. Replaced it with the tags set to always.

Replaced regex_replace with basename for trimming the /dev/.

Removed the debug tasks.  I intended to leave the debug to aid in debugging.

Fixed up the description in the name (use local case alphabets).