Add option to use jemalloc for Debian & Red Hat
Use "ceph_jemalloc_enabled" to enable jemalloc as the default memory
allocator for Ceph installs. This defaults to false, so no change to
existing deploys, but can be set to install and configure jemalloc as
the default memory allocator.

If the LD_PRELOAD line is changed in /etc/default/ceph, for Debian, or
/etc/sysconfig/ceph, for Red Hat, it will cause a restart of the ceph
services so that jemalloc is in use.